+++ src/commands/ls.js
@@ -1,13 +1,12 @@
 const chalk = require("chalk");
 const fs = require("fs");
 
-const _parseDoubleQuotes = require("../functions/parseQuotes");
 const _fatalError = require("../functions/fatalError");
-const _caseSensitivePath = require("../functions/caseSensitivePath");
 
 const Errors = require("../classes/Errors");
 const Checks = require("../classes/Checks");
 const Verbose = require("../classes/Verbose");
+const PathUtil = require("../classes/PathUtil");
 
 /**
  * Log the directory contents.
@@ -104,7 +103,7 @@ const ls = (dir = `"${process.cwd()}"`, ...args) => {
     if (dir === "-s") dir = process.cwd();
 
     Verbose.parseQuotes();
-    dir = _caseSensitivePath(_parseDoubleQuotes([dir, ...args])[0]);
+    dir = PathUtil.all([dir, ...args], { convertAbsolute: false });
 
     Verbose.initChecker();
     const dirChk = new Checks(dir);